Refactor event pitch crew with enhanced architecture and error handling
